Anna converter from xml to json is pending
Current converter assumed that new diameter development
is going to use the direct conversion from xml to json,
but this is not true because RFC6733 simplifies some
things and we will refactor dictionaries and message
formats.

So, we  prepare the intermediate function to adapt the
direct json to the real json once specified.